requirements.txt
# Core
python-dotenv>=1.0.0
pyyaml>=6.0

# Dashboard
streamlit>=1.40.0

# Google APIs
google-ads>=25.0.0
google-api-python-client>=2.150.0
google-auth>=2.35.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
gspread>=6.1.0

# Shopify
ShopifyAPI>=12.6.0

# OpenAI
openai>=1.50.0

# SerpAPI
google-search-results>=2.4.2

# Scheduling
apscheduler>=3.10.4

# Data processing
pandas>=2.2.0

# HTTP
requests>=2.32.0
httpx>=0.27.0

# Image processing
Pillow>=10.4.0

# NLP (for competition analysis)
scikit-learn>=1.5.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0
